+++ src/AreaChart/AreaChart.story.tsx
@@ -10,7 +10,8 @@ import {
   singleDateBigIntData,
   randomNumber,
   nonZeroDateData,
-  longMultiDateData
+  longMultiDateData,
+  binnedDateData
 } from '../../demo';
 import { AreaChart } from './AreaChart';
 import { StackedNormalizedAreaChart } from './StackedNormalizedAreaChart';
@@ -25,13 +26,21 @@ import {
   PointSeries
 } from './AreaSeries';
 import { GridlineSeries, Gridline } from '../common/Gridline';
-import { LinearXAxis, LinearXAxisTickSeries } from '../common/Axis/LinearAxis';
+import {
+  LinearXAxis,
+  LinearXAxisTickSeries,
+  LinearYAxis,
+  LinearYAxisTickSeries,
+  LinearYAxisTickLabel,
+  LinearXAxisTickLabel
+} from '../common/Axis/LinearAxis';
 import { ScatterPoint } from '../ScatterPlot';
 import { symbol, symbolStar } from 'd3-shape';
 import { Gradient, GradientStop } from '../common/Gradient';
 import { Stripes } from '../common/Mask';
 import { ChartDataShape } from '../common/data';
 import { schemes } from '../common/color';
+import { getYScale, getXScale } from '../common/scales';
 
 storiesOf('Demos|Area Chart/Single Series', module)
   .add(
@@ -342,6 +351,126 @@ storiesOf('Demos|Area Chart/Circle Series', module)
       }
     />
   ));
+
+storiesOf('Demos|Area Chart/Axis', module)
+  .add('Top + Bottom Axis', () => {
+    const scale = getXScale({
+      type: 'category',
+      width: 450,
+      data: [
+        {
+          key: 'Before',
+          data: 0,
+          x: 'Before'
+        },
+        {
+          key: 'After',
+          data: 0,
+          x: 'After'
+        }
+      ]
+    });
+
+    return (
+      <AreaChart
+        width={450}
+        height={200}
+        margins={0}
+        data={singleDateData}
+        xAxis={
+          <LinearXAxis
+            type="time"
+            orientation="horizontal"
+            position="end"
+            axisLine={null}
+            tickSeries={
+              <LinearXAxisTickSeries
+                line={null}
+                label={<LinearXAxisTickLabel padding={5} position="end" />}
+              />
+            }
+          />
+        }
+        secondaryAxis={[
+          <LinearXAxis
+            type="category"
+            orientation="horizontal"
+            position="start"
+            scale={scale}
+            axisLine={null}
+            tickSeries={
+              <LinearXAxisTickSeries
+                line={null}
+                label={<LinearXAxisTickLabel padding={20} position="start" />}
+              />
+            }
+          />
+        ]}
+        yAxis={<LinearYAxis type="value" axisLine={null} />}
+      />
+    );
+  })
+  .add('Left + Right Axis', () => {
+    const scale = getYScale({
+      type: 'category',
+      height: 200,
+      data: [
+        {
+          key: 'Low',
+          data: 0,
+          y: 'Low'
+        },
+        {
+          key: 'High',
+          data: 0,
+          y: 'High'
+        }
+      ]
+    });
+
+    return (
+      <AreaChart
+        width={450}
+        height={200}
+        margins={0}
+        data={singleDateData}
+        yAxis={
+          <LinearYAxis
+            position="end"
+            axisLine={null}
+            tickSeries={
+              <LinearYAxisTickSeries
+                line={null}
+                label={<LinearYAxisTickLabel padding={5} position="end" />}
+              />
+            }
+          />
+        }
+        secondaryAxis={[
+          <LinearYAxis
+            type="category"
+            position="start"
+            axisLine={null}
+            scale={scale}
+            tickSeries={
+              <LinearYAxisTickSeries
+                line={null}
+                label={
+                  <LinearYAxisTickLabel
+                    padding={20}
+                    position="start"
+                    rotation={270}
+                    align="start"
+                  />
+                }
+              />
+            }
+          />
+        ]}
+        xAxis={<LinearXAxis type="time" axisLine={null} />}
+      />
+    );
+  });
 
 const LiveUpdatingStory = () => {
   const [data, setData] = useState(multiDateData.map(d => ({ ...d })));