+++ src/ScatterPlot/ScatterPlot.story.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react';
+import React, { Fragment, useState } from 'react';
 import { storiesOf } from '@storybook/react';
 import { object, color, number } from '@storybook/addon-knobs';
 
@@ -10,17 +10,21 @@ import {
   signalStageData,
   signalStages
 } from '../../demo/signals';
-import { randomNumber } from '../../demo';
+import { randomNumber, singleDateData } from '../../demo';
 import { range } from 'd3-array';
 import { GridlineSeries, Gridline, GridStripe } from '../common/Gridline';
 import { ScatterSeries, ScatterPoint } from './ScatterSeries';
 import {
   LinearYAxis,
   LinearYAxisTickSeries,
-  LinearYAxisTickLabel
+  LinearYAxisTickLabel,
+  LinearXAxis,
+  LinearXAxisTickSeries,
+  LinearXAxisTickLabel
 } from '../common/Axis/LinearAxis';
 import { symbolStar, symbol } from 'd3-shape';
 import { schemes } from '../common/color';
+import { getYScale, getXScale } from '../common/scales';
 
 storiesOf('Demos|Scatter Plot/Linear', module)
   .add(
@@ -149,47 +153,164 @@ storiesOf('Demos|Scatter Plot/Linear', module)
   ))
   .add('Live Update', () => <BubbleChartLiveUpdate />);
 
-class BubbleChartLiveUpdate extends React.Component<any, any> {
-  constructor(props) {
-    super(props);
-    this.state = {
-      data: largeSignalChartData.map(d => ({ ...d }))
-    };
-  }
+storiesOf('Demos|Scatter Plot/Axis', module)
+  .add('Top + Bottom Axis', () => {
+    const scale = getXScale({
+      type: 'category',
+      width: 450,
+      data: [
+        {
+          key: 'Before',
+          data: 0,
+          x: 'Before'
+        },
+        {
+          key: 'After',
+          data: 0,
+          x: 'After'
+        }
+      ],
+      isMultiSeries: false,
+      isDiverging: true
+    });
 
-  updateData = () => {
-    const data = this.state.data.map(item => {
+    return (
+      <ScatterPlot
+        width={450}
+        height={200}
+        margins={0}
+        data={singleDateData}
+        xAxis={
+          <LinearXAxis
+            type="time"
+            orientation="horizontal"
+            position="end"
+            axisLine={null}
+            tickSeries={
+              <LinearXAxisTickSeries
+                line={null}
+                label={<LinearXAxisTickLabel padding={5} position="end" />}
+              />
+            }
+          />
+        }
+        secondaryAxis={[
+          <LinearXAxis
+            type="category"
+            orientation="horizontal"
+            position="start"
+            scale={scale}
+            axisLine={null}
+            tickSeries={
+              <LinearXAxisTickSeries
+                line={null}
+                label={<LinearXAxisTickLabel padding={20} position="start" />}
+              />
+            }
+          />
+        ]}
+        yAxis={<LinearYAxis type="value" axisLine={null} />}
+      />
+    );
+  })
+  .add('Left + Right Axis', () => {
+    const scale = getYScale({
+      type: 'category',
+      height: 200,
+      data: [
+        {
+          key: 'Low',
+          data: 0,
+          y: 'Low'
+        },
+        {
+          key: 'High',
+          data: 0,
+          y: 'High'
+        }
+      ],
+      isMultiSeries: false,
+      isDiverging: true
+    });
+
+    return (
+      <ScatterPlot
+        width={450}
+        height={200}
+        margins={0}
+        data={singleDateData}
+        yAxis={
+          <LinearYAxis
+            position="end"
+            axisLine={null}
+            tickSeries={
+              <LinearYAxisTickSeries
+                line={null}
+                label={<LinearYAxisTickLabel padding={5} position="end" />}
+              />
+            }
+          />
+        }
+        secondaryAxis={[
+          <LinearYAxis
+            type="category"
+            position="start"
+            axisLine={null}
+            scale={scale}
+            tickSeries={
+              <LinearYAxisTickSeries
+                line={null}
+                label={
+                  <LinearYAxisTickLabel
+                    padding={20}
+                    position="start"
+                    rotation={270}
+                    align="start"
+                  />
+                }
+              />
+            }
+          />
+        ]}
+        xAxis={<LinearXAxis type="time" axisLine={null} />}
+      />
+    );
+  });
+
+const BubbleChartLiveUpdate = () => {
+  const [data, setData] = useState(largeSignalChartData.map(d => ({ ...d })));
+
+  const updateData = () => {
+    const d = data.map(item => {
       item.data = randomNumber(1, 100);
       return { ...item };
     });
 
-    this.setState({ data });
+    setData(d);
   };
 
-  render() {
-    return (
-      <Fragment>
-        <ScatterPlot
-          height={400}
-          width={750}
-          data={this.state.data}
-          margins={20}
-          series={
-            <ScatterSeries
-              point={
-                <ScatterPoint
-                  color="rgba(45, 96, 232, .8)"
-                  size={v => {
-                    return v.metadata.severity + 5;
-                  }}
-                />
-              }
-            />
-          }
-        />
-        <br />
-        <button onClick={this.updateData}>Update</button>
-      </Fragment>
-    );
-  }
-}
+  return (
+    <Fragment>
+      <ScatterPlot
+        height={400}
+        width={750}
+        data={data}
+        margins={20}
+        series={
+          <ScatterSeries
+            point={
+              <ScatterPoint
+                color="rgba(45, 96, 232, .8)"
+                size={v => {
+                  return v.metadata.severity + 5;
+                }}
+              />
+            }
+          />
+        }
+      />
+      <br />
+      <button onClick={updateData}>Update</button>
+    </Fragment>
+  );
+};