+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-# Master - 10/7/19
+# 5.1.0 - 10/7/19
+- [Feature] Add ability to pass secondary x/y axis to bar/area/scatter/line/heat
 - [Feature] Add ability to pass in data to map and decouple it from build
 - [Chore] Start adding new docs